LPD-22762 Support download CSV report in the Individuals > Known Individuals > Overview
